workaround for hoedown's NULL-as-empty-buffer
The underlying Hoedown library sometimes passes `NULL` to render
callbacks to signify that the buffer is empty. This changes
`Buffer::from` so that, if the pointer is `NULL`, it creates a 0-sized
`Buffer` and returns that instead.

This should be safe because the semantics of `from` are that it returns
a read-only `Buffer`.

Also added a regression test.