Major refactorings
This implements Deref for Buffer to access the underlying bytes.

It also implements many conversion traits for Buffer and Markdown

The default method of constructing a Markdown instance is now using
a string slice, but the `from` method allows constructing from other
things such as bytes and other buffers. The previous constructor for
initializing from a Reader is now named `read_from`.

AsRef and AsMut were implemented for Buffer to replace the previous
as_slice and as_mut_slice methods, following std conventions. This might
require explicit bindings with annotations to aid type inference, which
may be alleviated in the future with type ascription.

Buffer's as_str was renamed to to_str, in line with the convention
established by the std (e.g. PathBuf::to_str).