defined the Wrapper trait and wrap! macro
The `Wrapper` trait facilitates the process of wrapping an existing
renderer and overriding certain behavior, while passing-through the rest
of the non-overridden behavior to the wrapped renderer.

The trait contains an associated type specifying the wrapped type, as
well as a `base()` method which yields a mutable reference to the
wrapped renderer. Default method implementations exist for each render
trait, which simply delegate to the wrapped renderer's respective
methods.

The `wrap!` macro takes the type of the wrapper as an argument and
generates an implementation of Render for it which simply uses UFCS to
pass-through the parameters to the `Wrapper` methods

Here is an `HtmlWrapper` type which wraps an html renderer. An
implementation of `Wrapper` is provided so that the user may override
specific behavior, for example the emphasis handler. The `wrap!` macro
then generates a `Render` implementation which leverages the `Wrapper`
implementation.

``` rust
pub struct HtmlWrapper {
    html: renderer::html::Html,
}

wrap!(HtmlWrapper);

impl Wrapper for HtmlWrapper {
    type Base = renderer::html::Html;

    #[inline(always)]
    fn base(&mut self) -> &mut renderer::html::Html {
        &mut self.html
    }

    fn emphasis(&mut self, ob: &mut Buffer, content: &Buffer) -> bool {
        write!(ob, "~~{}~~", content.to_str().unwrap()).unwrap();

        true
    }
}

fn test_html_wrapper() {
    let doc = Markdown::new(
"something _EMPHASIZED_ **hehe**");
    let html_renderer = HtmlWrapper {
        html: renderer::html::Html::new(renderer::html::Flags::empty(), 0),
    };

    let output = doc.render_to_buffer(html_renderer);

    assert_eq!(
        output.to_str().unwrap(),
        "<p>something ~~EMPHASIZED~~ <strong>hehe</strong></p>\n");
}
```