Modified build script to not use make.
The build script doesn't work reliably on Windows, which needs either a
different Makefile and VC++ or to be run inside MSYS.  Both of these are
a pain.

Instead, I've modified it to just call into `gcc` directly.  Although this
is rather more brittle (it'll break as soon as the source layout of
libhoedown changes), it's much simpler.