better handle null buffers
Sorry about the trouble, but it's kind of a tricky situation with
respect to how libhoedown uses a nullptr hoedown_buffer to represent an
optional hoedown_buffer, something that's pretty normal in C of course.

Currently a signature looks like:

``` rust
fn paragraph(&mut self, output: &mut Buffer, content: &Buffer)
```

However, `content` as received from libhoedown may be nullptr. So the
parameter should perhaps be changed to Option<&Buffer> or
&Option<Buffer>.

What this fix does for now is create a 0-sized hoedown_buffer which
works like any other, but a new `is_null` internal flag specifies
whether, when converting to a raw pointer of a hoedown_buffer which
libhoedown needs, it may convert to ptr::null() instead, as is
contention in libhoedown. Otherwise there are no null-pointers used in
Buffer, it's all at the point in which a raw pointer representation of
the Buffer is required, where a nullptr may be returned to libhoedown,
who knows what to do with it. Specifically this occurs in the
as_raw_ptr() function.

The reason an `is_null` flag is kept instead of just storing the nullptr
that libhoedown provided is that this way the Buffer acts as an "empty
buffer", making it usable like any other.

I'm honestly not very happy with this solution but it works for now, and
it's backwards-compatible for the most part. However, Buffer.as_ptr was
renamed to Buffer.data_ptr which makes more sense I think.