remove Buffer's Read implementation
The previous implementation was incorrect because it essentially
recreated a slice of the Buffer on each call, meaning that read progress
information was lost across calls.

The implementation has been removed since the implementation was very
basic anyways.

Going forward, one can Read from a Buffer by first getting a slice of
the Buffer and then using _its_ Read implementation. See the regression
test in this commit for an example of using read_to_end().

Fixes #6